Encoders are working! We have both standard and PID based commands.
PID requires that the average encoder position be returned as a double.

+++ EncoderTestCommand/src/org/usfirst/frc/team3166/robot/commands/DriveStraight.java
@@ -11,7 +11,7 @@ import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
  */
 public class DriveStraight extends Command {
 	private double m_motorSpeed;
-	private int encoderPos;
+	private double encoderPos;
 	public int m_distance;
 	
     public DriveStraight(int distance, double motorSpeed) {
